add the --force flag to delete command

+++ internal/cmd/postgresflex/instance/delete/delete.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stackitcloud/stackit-cli/internal/pkg/confirm"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/errors"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/examples"
+	"github.com/stackitcloud/stackit-cli/internal/pkg/flags"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/globalflags"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/services/postgresflex/client"
 	postgresflexUtils "github.com/stackitcloud/stackit-cli/internal/pkg/services/postgresflex/utils"
@@ -21,23 +22,35 @@ import (
 
 const (
 	instanceIdArg = "INSTANCE_ID"
+
+	forceDeleteFlag = "force"
+
+	deletedStatus = "Deleted"
 )
 
 type inputModel struct {
 	*globalflags.GlobalFlagModel
-	InstanceId string
+	InstanceId  string
+	ForceDelete bool
 }
 
 func NewCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   fmt.Sprintf("delete %s", instanceIdArg),
 		Short: "Deletes a PostgreSQL Flex instance",
-		Long:  "Deletes a PostgreSQL Flex instance.",
-		Args:  args.SingleArg(instanceIdArg, utils.ValidateUUID),
+		Long: fmt.Sprintf("%s\n%s\n%s",
+			"Deletes a PostgreSQL Flex instance.",
+			"By default, instances will be kept in a delayed deleted state for 7 days before being permanently deleted.",
+			"Use the --force flag to force the deletion of a delayed deleted instance.",
+		),
+		Args: args.SingleArg(instanceIdArg, utils.ValidateUUID),
 		Example: examples.Build(
 			examples.NewExample(
 				`Delete a PostgreSQL Flex instance with ID "xxx"`,
 				"$ stackit postgresflex instance delete xxx"),
+			examples.NewExample(
+				`Force the deletion of a delayed deleted PostgreSQL Flex instance with ID "xxx"`,
+				"$ stackit postgresflex instance delete xxx --force"),
 		),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			ctx := context.Background()
@@ -65,22 +78,44 @@ func NewCmd() *cobra.Command {
 				}
 			}
 
-			// Call API
-			req := buildRequest(ctx, model, apiClient)
-			err = req.Execute()
+			instanceStatus, err := postgresflexUtils.GetInstanceStatus(ctx, apiClient, model.ProjectId, model.InstanceId)
 			if err != nil {
-				return fmt.Errorf("delete PostgreSQL Flex instance: %w", err)
+				return fmt.Errorf("get PostgreSQL Flex instance status: %w", err)
 			}
 
-			// Wait for async operation, if async mode not enabled
-			if !model.Async {
-				s := spinner.New(cmd)
-				s.Start("Deleting instance")
-				_, err = wait.DeleteInstanceWaitHandler(ctx, apiClient, model.ProjectId, model.InstanceId).WaitWithContext(ctx)
+			if instanceStatus == deletedStatus && !model.ForceDelete {
+				return fmt.Errorf("instance %q is already deleted, use --force to force the deletion of a delayed deleted instance", instanceLabel)
+			}
+
+			if instanceStatus != deletedStatus {
+				// Call API
+				delReq := buildDeleteRequest(ctx, model, apiClient)
+				err = delReq.Execute()
 				if err != nil {
-					return fmt.Errorf("wait for PostgreSQL Flex instance deletion: %w", err)
+					return fmt.Errorf("delete PostgreSQL Flex instance: %w", err)
 				}
-				s.Stop()
+
+				// Wait for async operation, if async mode not enabled
+				if !model.Async {
+					s := spinner.New(cmd)
+					s.Start("Deleting instance")
+					_, err = wait.DeleteInstanceWaitHandler(ctx, apiClient, model.ProjectId, model.InstanceId).WaitWithContext(ctx)
+					if err != nil {
+						return fmt.Errorf("wait for PostgreSQL Flex instance deletion: %w", err)
+					}
+					s.Stop()
+				}
+			}
+
+			if model.ForceDelete {
+				// Call API
+				forceDelReq := buildForceDeleteRequest(ctx, model, apiClient)
+				err = forceDelReq.Execute()
+				if err != nil {
+					return fmt.Errorf("force delete PostgreSQL Flex instance: %w", err)
+				}
+
+				// TODO: Wait for async operation, if async mode not enabled
 			}
 
 			operationState := "Deleted"
@@ -91,7 +126,12 @@ func NewCmd() *cobra.Command {
 			return nil
 		},
 	}
+	configureFlags(cmd)
 	return cmd
+}
+
+func configureFlags(cmd *cobra.Command) {
+	cmd.Flags().BoolP(forceDeleteFlag, "f", false, "Force deletion of a delayed deleted instance")
 }
 
 func parseInput(cmd *cobra.Command, inputArgs []string) (*inputModel, error) {
@@ -105,10 +145,16 @@ func parseInput(cmd *cobra.Command, inputArgs []string) (*inputModel, error) {
 	return &inputModel{
 		GlobalFlagModel: globalFlags,
 		InstanceId:      instanceId,
+		ForceDelete:     flags.FlagToBoolValue(cmd, forceDeleteFlag),
 	}, nil
 }
 
-func buildRequest(ctx context.Context, model *inputModel, apiClient *postgresflex.APIClient) postgresflex.ApiDeleteInstanceRequest {
+func buildDeleteRequest(ctx context.Context, model *inputModel, apiClient *postgresflex.APIClient) postgresflex.ApiDeleteInstanceRequest {
 	req := apiClient.DeleteInstance(ctx, model.ProjectId, model.InstanceId)
+	return req
+}
+
+func buildForceDeleteRequest(ctx context.Context, model *inputModel, apiClient *postgresflex.APIClient) postgresflex.ApiForceDeleteInstanceRequest {
+	req := apiClient.ForceDeleteInstance(ctx, model.ProjectId, model.InstanceId)
 	return req
 }
